metly_backend/pyproject.toml
[project]
name = "metly_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Rasmus Stenstrøm Johansen",email = "rasmus@metly.dk"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "lightgbm (>=4.5.0,<5.0.0)",
    "email-validator (>=2.0.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)"
]



[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
